# Filter plugin records count and record size metrics