--- a/lib/fluent/plugin/filter.rb
+++ b/lib/fluent/plugin/filter.rb
@@ -38,6 +38,13 @@ module Fluent
         @emit_records = 0
         @emit_size = 0
         @counter_mutex = Mutex.new
+        @enable_size_metrics = false
+      end
+
+      def configure(conf)
+        super
+
+        @enable_size_metrics = !!system_config.enable_size_metrics
       end
 
       def statistics
@@ -52,7 +59,7 @@ module Fluent
       def measure_metrics(es)
         @counter_mutex.synchronize do
           @emit_records += es.size
-          @emit_size += es.to_msgpack_stream.bytesize
+          @emit_size += es.to_msgpack_stream.bytesize if @enable_size_metrics
         end
       end