--- a/lib/fluent/plugin/in_monitor_agent.rb
+++ b/lib/fluent/plugin/in_monitor_agent.rb
@@ -336,6 +336,7 @@ module Fluent::Plugin
 
       if pe.respond_to?(:statistics)
         obj.merge!(pe.statistics.dig('output') || {})
+        obj.merge!(pe.statistics.dig('filter') || {})
         obj.merge!(pe.statistics.dig('input') || {})
       end