<!--
Thank you for contributing to Fluentd!
Your commits need to follow DCO: https://probot.github.io/apps/dco/
And please provide the following information to help us make the most of your pull request:
-->

**Which issue(s) this PR fixes**: 
None

**What this PR does / why we need it**: 
Currently, Fluentd filter metrics do not have emit size and emitted record counts for calculating flow rate (per sec or per hour).
Fluentd filter should have emit_records and emit_size metrics.

But, currently, I disabled size metrics by default for performance concerns.

**Docs Changes**:
Needed.

**Release Note**: 
Same as title.