@cosmo0920 2021-07-28
I measured performance with micro benchmark and the following patch:

```diff
diff --git a/lib/fluent/plugin/filter.rb b/lib/fluent/plugin/filter.rb
index 909ba82c..282243ad 100644
--- a/lib/fluent/plugin/filter.rb
+++ b/lib/fluent/plugin/filter.rb
@@ -31,6 +31,7 @@ module Fluent
       helpers_internal :event_emitter
 
       attr_reader :has_filter_with_time
+      attr_accessor :enable_size_metrics
 
       def initialize
         super
```

```ruby
#!/usr/bin/env ruby

require 'benchmark/ips'

require_relative 'test/helper'
require 'fluent/plugin/filter'
require 'fluent/event'
require 'fluent/engine'
require 'fluent/event_router'

class DummyStreamFilter < Fluent::Plugin::Filter
  def filter_stream(tag, es)
    es
  end
end

class DummyFilter < Fluent::Plugin::Filter
  def filter(tag, time, record)
    record
  end
end

def create_filter_stream(router)
  f = DummyStreamFilter.new
  f.context_router = router
  f.configure(config_element())
  f
end

def create_filter(router)
  f = DummyFilter.new
  f.context_router = router
  f.configure(config_element())
  f
end

router = Fluent::EventRouter.new(Fluent::NoMatchMatch.new($log), self)

fs1 = create_filter_stream(router)
fs1.enable_size_metrics = false

fs2 = create_filter_stream(router)

f1 = create_filter(router)
f1.enable_size_metrics = false

f2 = create_filter(router)

t = event_time()
es = Fluent::ArrayEventStream.new([ [t, {"key" => "value1"}], [t, {"key" => "value2"}] ])

Benchmark.ips do |x|
  x.report("before -- filter stream") do
    fs1.filter_stream('tag', es)
  end

  x.report("after -- filter stream") do
    fs2.filter_stream('tag', es)
  end

  x.report("before -- filter") do
    f1.filter('tag', t, {"key" => "value1"})
  end

  x.report("after -- filter") do
    f2.filter('tag', t, {"key" => "value1"})
  end
end
```

## Results

### master branch

```console
$ bundle exec ruby bench_filter_metrics.rb
Warming up --------------------------------------
before -- filter stream
                       596.528k i/100ms
after -- filter stream
                       593.494k i/100ms
    before -- filter   384.501k i/100ms
     after -- filter   392.635k i/100ms
Calculating -------------------------------------
before -- filter stream
                          5.919M (± 1.0%) i/s -     29.826M in   5.039988s
after -- filter stream
                          5.819M (± 3.3%) i/s -     29.081M in   5.004079s
    before -- filter      3.889M (± 1.8%) i/s -     19.610M in   5.044586s
     after -- filter      3.871M (± 1.5%) i/s -     19.632M in   5.072232s
```

### This branch

```console
$ bundle exec ruby bench_filter_metrics.rb
Warming up --------------------------------------
before -- filter stream
                       593.929k i/100ms
after -- filter stream
                       594.544k i/100ms
    before -- filter   390.394k i/100ms
     after -- filter   387.836k i/100ms
Calculating -------------------------------------
before -- filter stream
                          5.704M (± 5.8%) i/s -     28.509M in   5.019498s
after -- filter stream
                          5.885M (± 2.1%) i/s -     29.727M in   5.053743s
    before -- filter      3.920M (± 1.1%) i/s -     19.910M in   5.080193s
     after -- filter      3.905M (± 1.1%) i/s -     19.780M in   5.066421s
```

---

No significant performance degradation is observed.